+++ org-mode/lib/Data/Org.hs
@@ -5,7 +5,7 @@
 
 -- |
 -- Module    : Data.Org
--- Copyright : (c) Colin Woodbury, 2020
+-- Copyright : (c) Colin Woodbury, 2020 - 2021
 -- License   : BSD3
 -- Maintainer: Colin Woodbury <colin@fosskers.ca>
 --
@@ -19,7 +19,9 @@ module Data.Org
   , emptyOrgFile
   , OrgDoc(..)
   , emptyDoc
+  , allDocTags
   , Section(..)
+  , allSectionTags
   , Block(..)
   , Words(..)
   , ListItems(..)
@@ -55,6 +57,8 @@ import           Data.List.NonEmpty (NonEmpty(..))
 import qualified Data.List.NonEmpty as NEL
 import qualified Data.Map.Strict as M
 import           Data.Semigroup (sconcat)
+import           Data.Set (Set)
+import qualified Data.Set as S
 import           Data.Text (Text)
 import qualified Data.Text as T
 import           Data.Void (Void)
@@ -102,6 +106,16 @@ data OrgDoc = OrgDoc
 emptyDoc :: OrgDoc
 emptyDoc = OrgDoc [] []
 
+-- | All unique section tags in the entire document.
+--
+-- Section tags appear on the same row as a header title, but right-aligned.
+--
+-- @
+-- * This is a Heading                :tag1:tag2:
+-- @
+allDocTags :: OrgDoc -> Set Text
+allDocTags = foldMap allSectionTags . docSections
+
 -- | Some logically distinct block of Org content.
 data Block
   = Quote Text
@@ -127,6 +141,10 @@ data Section = Section
   , sectionDoc     :: OrgDoc }
   deriving stock (Eq, Show, Generic)
   deriving anyclass (Hashable)
+
+-- | All unique tags with a section and its subsections.
+allSectionTags :: Section -> Set Text
+allSectionTags (Section _ sts doc) = S.fromList sts <> allDocTags doc
 
 -- | An org list constructed of @-@ characters.
 --
@@ -175,7 +193,6 @@ data Words
   | Strike Text
   | Link URL (Maybe Text)
   | Image URL
-  | Tags (NonEmpty Text)
   | Punct Char
   | Plain Text
   deriving stock (Eq, Show, Generic)
@@ -231,14 +248,14 @@ orgP' depth = L.lexeme space $ OrgDoc
       , try table
       , paragraph ]  -- TODO Paragraph needs to fail if it detects a heading.
 
--- | If a line stars with @*@ and a space, it is a `Section` heading.
+-- | If a line starts with @*@ and a space, it is a `Section` heading.
 heading :: Parser (T.Text, NonEmpty Words, [Text])
 heading = do
   stars <- someOf '*' <* char ' '
-  ws <- line '\n'
-  case nelUnsnoc ws of
-    (Tags ts, Just rest) -> pure (stars, rest, NEL.toList ts)
-    _                    -> pure (stars, ws, [])
+  (ws, mts) <- headerLine
+  case mts of
+    Nothing -> pure (stars, ws, [])
+    Just ts -> pure (stars, ws, NEL.toList ts)
 
 section :: Int -> Parser Section
 section depth = L.lexeme space $ do
@@ -332,8 +349,14 @@ paragraph = L.lexeme space $ do
   notFollowedBy heading
   Paragraph . sconcat <$> sepEndBy1 (line '\n') newline
 
+headerLine :: Parser (NonEmpty Words, Maybe (NonEmpty Text))
+headerLine = do
+  ws <- (wordChunk '\n' <* hspace) `someTill` lookAhead (void tags <|> void (char '\n') <|> eof)
+  ts <- optional tags
+  pure (ws, ts)
+
 line :: Char -> Parser (NonEmpty Words)
-line end = sepEndBy1 (wordChunk end) (manyOf ' ')
+line end = wordChunk end `sepEndBy1` manyOf ' '
 
 -- | RULES
 --
@@ -354,20 +377,20 @@ wordChunk end = choice
   , try $ Strike    <$> between (char '+') (char '+') (someTill' '+') <* pOrS
   , try image
   , try link
-  , try tags
   , try $ Punct     <$> oneOf punc
   , Plain           <$> takeWhile1P (Just "plain text") (\c -> c /= ' ' && c /= end) ]
   where
+    -- | Punctuation, space, or the end of the file.
     pOrS :: Parser ()
     pOrS = lookAhead $ void (oneOf $ end : ' ' : punc) <|> eof
 
 punc :: String
 punc = ".,!?():;'"
 
-tags :: Parser Words
+tags :: Parser (NonEmpty Text)
 tags = do
   void $ char ':'
-  Tags <$> (T.pack . NEL.toList <$> some letterChar) `sepEndBy1` char ':'
+  (T.pack . NEL.toList <$> some (alphaNumChar <|> char '_' <|> char '@')) `sepEndBy1` char ':'
 
 image :: Parser Words
 image = between (char '[') (char ']') $
@@ -479,9 +502,5 @@ prettyWords w = case w of
   Link (URL url) Nothing  -> "[[" <> url <> "]]"
   Link (URL url) (Just t) -> "[[" <> url <> "][" <> t <> "]]"
   Image (URL url)         -> "[[" <> url <> "]]"
-  Tags ts                 ->  ":" <> T.intercalate ":" (NEL.toList ts) <> ":"
   Punct c                 -> T.singleton c
   Plain t                 -> t
-
-nelUnsnoc :: NonEmpty a -> (a, Maybe (NonEmpty a))
-nelUnsnoc ne = (NEL.last ne, NEL.nonEmpty $ NEL.init ne)